+++ changelog.md
@@ -5,6 +5,7 @@ Upcoming release
 ================
 
 ## New features
+- Single case test for slope index
 - Smaller refinements